*More clean up on user authentication.

Closes #28

See merge request lunia-development/Lunia!47

+++ Info/Info/StageGroup/StageGroupManager.cpp
@@ -27,7 +27,7 @@ namespace Lunia{
 					std::vector< uint32 >::const_iterator it = HighPriorityStageGroups.begin();
 					for (auto count = HighPriorityStageGroups.size(); it != HighPriorityStageGroups.end(); ++it, --count)
 					{
-						stageGroupPriority[*it] = count;
+						stageGroupPriority[*it] = (int)count;
 					}
 
 					it = LowPriorityStageGroups.begin();
